Refactor : action creator 적용
상세 :
- ChattingManager, ClientManager, GameManager에 적용
- 오타 수정

close #391
